# services/server/src/app.ts
import { timingSafeEqual } from "node:crypto";

import Fastify, { type FastifyInstance, type FastifyServerOptions } from "fastify";

import { TRANSITION_REASONS, WORK_ITEM_PRIORITIES, WORK_ITEM_STATUSES, WORK_ITEM_TYPES } from "@missiongo/domain";

import { invalidInput, MissionGoError } from "./errors.js";
import { MissionGoStore } from "./store.js";
import { COMPONENT_KINDS, type ComponentKind } from "./types.js";

export interface BuildAppOptions {
  readonly databasePath?: string;
  readonly logger?: FastifyServerOptions["logger"];
  readonly adminToken?: string;
}

function objectBody(value: unknown): Record<string, unknown> {
  if (!value || typeof value !== "object" || Array.isArray(value)) throw invalidInput("A JSON object is required.");
  return value as Record<string, unknown>;
}

function stringField(body: Record<string, unknown>, field: string, required = true): string | undefined {
  const value = body[field];
  if (value === undefined && !required) return undefined;
  if (typeof value !== "string") throw invalidInput(`${field} must be a string.`);
  return value;
}

function stringArrayField(body: Record<string, unknown>, field: string): readonly string[] | undefined {
  const value = body[field];
  if (value === undefined) return undefined;
  if (!Array.isArray(value) || value.some((entry) => typeof entry !== "string")) {
    throw invalidInput(`${field} must be an array of strings.`);
  }
  return value as string[];
}

function enumField<T extends string>(
  body: Record<string, unknown>,
  field: string,
  choices: readonly T[],
  required = true,
): T | undefined {
  const value = body[field];
  if (value === undefined && !required) return undefined;
  if (typeof value !== "string" || !choices.includes(value as T)) {
    throw invalidInput(`${field} must be one of: ${choices.join(", ")}.`);
  }
  return value as T;
}

export function buildApp(options: BuildAppOptions = {}): FastifyInstance {
  const app = Fastify({ logger: options.logger ?? false });
  const store = new MissionGoStore(options.databasePath ?? ":memory:");

  app.decorate("missionGoStore", store);
  app.addHook("onClose", () => store.close());

  app.addHook("onRequest", async (request, reply) => {
    if (!request.url.startsWith("/api/v1/") || !options.adminToken) return;
    const authorization = request.headers.authorization;
    const suppliedToken = authorization?.startsWith("Bearer ") ? authorization.slice(7) : "";
    const expected = Buffer.from(options.adminToken);
    const supplied = Buffer.from(suppliedToken);
    const authenticated = expected.length === supplied.length && timingSafeEqual(expected, supplied);
    if (!authenticated) {
      return reply.status(401).send({
        type: "urn:missiongo:problem:authentication_required",
        title: "A valid bearer token is required.",
        status: 401,
        code: "authentication_required",
      });
    }
  });

  app.setErrorHandler((error, _request, reply) => {
    if (error instanceof MissionGoError) {
      return reply.status(error.statusCode).send({
        type: `urn:missiongo:problem:${error.code}`,
        title: error.message,
        status: error.statusCode,
        code: error.code,
      });
    }

    app.log.error(error);
    return reply.status(500).send({
      type: "urn:missiongo:problem:internal_error",
      title: "An unexpected error occurred.",
      status: 500,
      code: "internal_error",
    });
  });

  app.get("/health", async () => ({ status: "ok" }));

  app.get("/api/v1/products", async () => store.listProducts());

  app.post("/api/v1/products", async (request, reply) => {
    const body = objectBody(request.body);
    const product = store.createProduct({
      name: stringField(body, "name")!,
      keyPrefix: stringField(body, "keyPrefix")!,
    });
    return reply.status(201).send(product);
  });

  app.get("/api/v1/products/:productId/components", async (request) => {
    const { productId } = request.params as { productId: string };
    return store.listComponents(productId);
  });

  app.post("/api/v1/products/:productId/components", async (request, reply) => {
    const { productId } = request.params as { productId: string };
    const body = objectBody(request.body);
    const component = store.createComponent({
      productId,
      name: stringField(body, "name")!,
      kind: enumField(body, "kind", COMPONENT_KINDS)! as ComponentKind,
    });
    return reply.status(201).send(component);
  });

  app.get("/api/v1/items", async (request) => {
    const query = request.query as Record<string, unknown>;
    const productId = typeof query.productId === "string" ? query.productId : undefined;
    if (!productId) throw invalidInput("productId is required.");
    const limit = typeof query.limit === "string" ? Number(query.limit) : undefined;
    const beforeSequence = typeof query.beforeSequence === "string" ? Number(query.beforeSequence) : undefined;
    if (limit !== undefined && (!Number.isInteger(limit) || limit < 1 || limit > 100)) {
      throw invalidInput("limit must be an integer between 1 and 100.");
    }
    if (beforeSequence !== undefined && (!Number.isInteger(beforeSequence) || beforeSequence < 1)) {
      throw invalidInput("beforeSequence must be a positive integer.");
    }

    const items = store.listWorkItems({
      productId,
      ...(typeof query.status === "string" ? { status: query.status as never } : {}),
      ...(typeof query.type === "string" ? { type: query.type as never } : {}),
      ...(limit !== undefined ? { limit } : {}),
      ...(beforeSequence !== undefined ? { beforeSequence } : {}),
    });
    return { items };
  });

  app.post("/api/v1/items", async (request, reply) => {
    const body = objectBody(request.body);
    const environment = body.environment;
    if (environment !== undefined && (!environment || typeof environment !== "object" || Array.isArray(environment))) {
      throw invalidInput("environment must be an object.");
    }
    const item = store.createWorkItem({
      productId: stringField(body, "productId")!,
      ...(stringField(body, "sourceComponentId", false) ? { sourceComponentId: body.sourceComponentId as string } : {}),
      ...(stringArrayField(body, "affectedComponentIds")
        ? { affectedComponentIds: stringArrayField(body, "affectedComponentIds")! }
        : {}),
      ...(stringField(body, "areaId", false) ? { areaId: body.areaId as string } : {}),
      type: enumField(body, "type", WORK_ITEM_TYPES)!,
      priority: enumField(body, "priority", WORK_ITEM_PRIORITIES)!,
      title: stringField(body, "title")!,
      description: stringField(body, "description")!,
      ...(environment ? { environment: environment as never } : {}),
    });
    return reply.status(201).send(item);
  });

  app.get("/api/v1/items/:itemKey", async (request) => {
    const { itemKey } = request.params as { itemKey: string };
    return store.getWorkItem(itemKey);
  });

  app.patch("/api/v1/items/:itemKey", async (request) => {
    const { itemKey } = request.params as { itemKey: string };
    const body = objectBody(request.body);
    return store.updateWorkItem(itemKey, {
      ...(stringField(body, "title", false) !== undefined ? { title: body.title as string } : {}),
      ...(stringField(body, "description", false) !== undefined ? { description: body.description as string } : {}),
      ...(body.type !== undefined ? { type: enumField(body, "type", WORK_ITEM_TYPES)! } : {}),
      ...(body.priority !== undefined ? { priority: enumField(body, "priority", WORK_ITEM_PRIORITIES)! } : {}),
      ...(body.affectedComponentIds !== undefined
        ? { affectedComponentIds: stringArrayField(body, "affectedComponentIds")! }
        : {}),
    });
  });

  app.post("/api/v1/items/:itemKey/transitions", async (request) => {
    const { itemKey } = request.params as { itemKey: string };
    const body = objectBody(request.body);
    return store.transitionWorkItem({
      itemKey,
      to: enumField(body, "to", WORK_ITEM_STATUSES)!,
      actor: "human",
      reason: enumField(body, "reason", TRANSITION_REASONS)!,
      ...(stringField(body, "note", false) !== undefined ? { note: body.note as string } : {}),
    });
  });

  app.get("/api/v1/items/:itemKey/timeline", async (request) => {
    const { itemKey } = request.params as { itemKey: string };
    return { events: store.getTimeline(itemKey) };
  });

  return app;
}

declare module "fastify" {
  interface FastifyInstance {
    missionGoStore: MissionGoStore;
  }
}